+++ hpack/__init__.py
@@ -5,4 +5,7 @@ hpack
 
 HTTP/2 header encoding for Python.
 """
-__version__ = '1.1.0'
+from .hpack import Encoder, Decoder
+from .exceptions import HPACKError, HPACKDecodingError
+
+__version__ = '2.0.0'